fix: emit fan RPM as integers (no decimals)
Oven and board fan RPM are now int across SensorReadings, TraceSample and their DTOs; the simulator casts to int. RPM has no meaningful fractional part.

+++ src/ReflowOven.Application/Dtos/DiagnosticsDtos.cs
@@ -25,9 +25,9 @@ public sealed record DiagnosticsOverviewDto(
 /// <summary>Live sensor readings (Diagnóstico → Sensores), also pushed over the diagnostics hub.</summary>
 public sealed record SensorReadingsDto(
     double BoardTempC,
-    double BoardFanRpm,
+    int BoardFanRpm,
     double OvenTempC,
-    double OvenFanRpm,
+    int OvenFanRpm,
     double VoltageV,
     double CurrentA)
 {
